computer-vision/requirements.txt
get-mnist==1.0.1
joblib==1.1.0
numpy==1.21.2
pkg_resources==0.0.0
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
threadpoolctl==3.0.0
